backend/requirements.txt
# Web Framework
fastapi==0.100.0
uvicorn==0.22.0

# Web Scraping
playwright==1.36.0
beautifulsoup4==4.12.2
selenium==4.15.0
requests==2.31.0
lxml==4.9.3

# HTTP Clients
httpx==0.24.1
aiohttp==3.8.5

# Data Processing
pandas==2.0.3
pydantic==2.1.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
aiosqlite==0.19.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Configuration & Environment
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Monitoring & Logging
slowapi==0.1.9
structlog==23.1.0

# Proxy Management
aiohttp-proxy==0.1.2

# Utilities
python-dateutil==2.8.2
croniter==1.4.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
factory-boy==3.3.0

# Development
black==23.9.1
flake8==6.1.0
mypy==1.6.0
